Added parenthesis/delimiter matching.
Added facilities to match various kinds of MathJax delimiters as well as
bracket symbols that might not result in a syntax error if not matched.
There's room for improvement, and it might be better to limit purely
symbolic brackets to only match within a grouping, but that will require
more in-depth parsing.